newtonsoft json added

+++ Services/ParticipantService.cs
@@ -47,11 +47,12 @@ namespace SplitExpenses.Services
             return result;
         }
 
-      public bool  GetParticipantByMobile(int mobile)
+      public bool  GetParticipantByMobile(long mobile)
         {
             bool result;
             if (mobile.ToString().Length == 0)
                 throw new Exception("Invalid mobile number");
+
             var participant = _unitOfWork.Repository<Participant>().FindBy(p => p.Mobile == mobile).FirstOrDefault();
             if (participant == null)
                 result = false;
